requirements.txt
tweepy
bs4
pyuser-agent
tabula-py